refactor: extract capture backends and helpers into ocr_capture_backends
Move Protocols, capture helper functions, and 5 capture backend classes
(MssCaptureBackend, PyAutoGuiCaptureBackend, PrintWindowCaptureBackend,
DxcamCaptureBackend, Win32CaptureBackend) into ocr_capture_backends.py.
Update test monkeypatches for the moved symbols.

ocr_reader.py: 7219 -> 6144 lines.
